Refactor our admin client
